Introduce router and prepare files for question edit form
