+++ ui/src/index.js
@@ -5,11 +5,11 @@ import {ApolloProvider} from 'react-apollo';
 import AppConfig from './services/AppConfig';
 import GraphqlClient from './services/GraphqlClient';
 
-import App from './components/App';
+import Root from './Root';
 
 ReactDOM.render(
   <ApolloProvider client={GraphqlClient({port: AppConfig.port()})}>
-    <App />
+    <Root />
   </ApolloProvider>,
   AppConfig.root(),
 );
