+++ ui/src/pages/Questions/Question.jsx
@@ -2,9 +2,11 @@ import React, {Component} from 'react';
 import {graphql} from 'react-apollo';
 import gql from 'graphql-tag';
 import DocumentTitle from 'react-document-title';
+import {withRouter} from 'react-router';
 import {Container, Loader, Grid, Header, Dropdown, Menu} from 'semantic-ui-react';
 import Markdown from 'react-remarkable';
 
+import {urlFor} from '../../helpers/routing';
 import ApiClient from '../../services/ApiClient';
 
 import QueryResult from '../../components/QueryResult';
@@ -71,6 +73,11 @@ class QuestionContainer extends Component {
     });
   };
 
+  handleEdit = question => {
+    const questionPath = urlFor(question, ['id', 'title']);
+    this.props.history.push(`/questions/${questionPath}/edit`);
+  };
+
   render() {
     const {data: {loading, error, question}} = this.props;
 
@@ -110,7 +117,7 @@ class QuestionContainer extends Component {
                 </Menu.Item>
               )}
               <Menu.Menu position="right">
-                <Menu.Item icon="edit" onClick={this.handleEdit} />
+                <Menu.Item icon="edit" onClick={() => this.handleEdit(question)} />
                 <Dropdown item icon="download">
                   <Dropdown.Menu>
                     <Dropdown.Item onClick={this.exportPNG}>PNG</Dropdown.Item>
@@ -144,8 +151,10 @@ const Query = gql`
   }
 `;
 
-const Question = graphql(Query, {
-  options: ({id}) => ({variables: {id}}),
-})(QuestionContainer);
+const Question = withRouter(
+  graphql(Query, {
+    options: ({id}) => ({variables: {id}}),
+  })(QuestionContainer),
+);
 
 export default Question;
