+++ ui/src/pages/Questions/QuestionEdit.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const QuestionEdit = () => {
+  return <div>WIP</div>;
+};
+
+export default QuestionEdit;
